pyproject.toml
[project]
name = "ascogu26-monitor"
version = "0.1.0"
description = "ASCO GU 2026 Twitter/X Monitoring Pipeline"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "tenacity>=8.0",
    "streamlit>=1.30",
    "pandas>=2.0",
    "plotly>=5.18",
    "scikit-learn>=1.3",
    "anthropic>=0.40",
]

[project.optional-dependencies]
bertopic = [
    "bertopic>=0.16",
    "sentence-transformers>=2.2",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "responses>=0.24",
]
